+++ config/routes.js
@@ -1,5 +1,6 @@
 const router = require('express').Router();
 const renderer = require('../app/middleware/renderer');
+const selectionValidator = require('../app/middleware/selectionValidator');
 
 router.get(
   '/',
@@ -9,6 +10,12 @@ router.get(
 router.get(
   '/symptoms',
   renderer.symptoms
+);
+
+router.get(
+  '/recommend',
+  selectionValidator,
+  renderer.recommend
 );
 
 module.exports = router;