+++ config/routes.js
@@ -1,6 +1,7 @@
 const router = require('express').Router();
 const renderer = require('../app/middleware/renderer');
 const selectionValidatorSymptoms = require('../app/middleware/selectionValidatorSymptoms');
+const selectionValidatorAge = require('../app/middleware/selectionValidatorAge');
 
 router.get(
   '/',
@@ -21,6 +22,12 @@ router.get(
 router.get(
   '/age',
   renderer.age
+);
+
+router.get(
+  '/choose',
+  selectionValidatorAge,
+  renderer.choose
 );
 
 module.exports = router;