+++ test/integration/start-page.js
@@ -8,20 +8,27 @@ const expect = chai.expect;
 
 chai.use(chaiHttp);
 
-describe('Search page', () => {
-  describe('page title', () => {
-    it('should be \'Find a chlamydia test - NHS.UK\'', async () => {
-      const res = await chai.request(server).get(`${constants.SITE_ROOT}`);
+describe('Start page', () => {
+  it('page title should be \'Find a chlamydia test - NHS.UK\'', async () => {
+    const res = await chai.request(server).get(`${constants.SITE_ROOT}`);
 
-      const $ = cheerio.load(res.text);
+    const $ = cheerio.load(res.text);
 
-      expect($('title').text()).to.equal('Find a chlamydia test - NHS.UK');
-      expect($('.local-header--title--question').text()).to.equal('Find a chlamydia test');
-    });
+    expect($('title').text()).to.equal('Find a chlamydia test - NHS.UK');
+    expect($('.local-header--title--question').text()).to.equal('Find a chlamydia test');
+  });
+
+  it('should link to the \'Symptoms\' page', async () => {
+    const res = await chai.request(server).get(`${constants.SITE_ROOT}`);
+    const $ = cheerio.load(res.text);
+    const symptomsPage = `${constants.SITE_ROOT}/symptoms`;
+
+    expect($('.start-button').attr('href'))
+      .to.equal(symptomsPage);
   });
 
   describe('return to Choices services', () => {
-    it('should have a link back to the Choices \'Services near you\'', async () => {
+    it('the breadcrumb should have a link back to the Choices \'Services near you\'', async () => {
       const res = await chai.request(server).get(`${constants.SITE_ROOT}`);
 
       const $ = cheerio.load(res.text);
@@ -30,13 +37,13 @@ describe('Search page', () => {
         .to.equal('https://www.nhs.uk/service-search');
     });
 
-    it('should have a link back to the Choices service search', async () => {
+    it('the page should have a link back to the Choices service search', async () => {
       const res = await chai.request(server).get(`${constants.SITE_ROOT}`);
 
       const $ = cheerio.load(res.text);
 
       expect($('.back-to-choices').attr('href'))
-        .to.equal('https://www.nhs.uk/Service-Search');
+        .to.equal('https://www.nhs.uk/service-search');
     });
   });
 });