+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+0.3.0 / TBA
+=======
+- Error pages
+- Symptoms page
+- Age page
+- Form error handling
+
 0.2.0 / 2018-03-01
 =======
 - Upgrade python in Alpine to 2.7.14-r0